New Changes
1. Added a Cpp file that contains the object oriented version of the
base code and is also responsible for the maintenance of the overall
outlook of the application. Still need to figure out the main problem
of adding data to end of file using ios::app for appending data to the
end of the file once we have created the password file.